feat(esco-content-menu): support sort by custom portlet order
if an "escoMenuOrder" parameter is provided that will be used to sort
the element.
If no custom order is provided it will sort by the title in the browsers
current locale.
When there is a mix of custom order and title order, custom order will
be moved in front of title order.